requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Migrate==4.0.7
alembic==1.14.0
python-dotenv==1.0.1
gunicorn
passlib[bcrypt]==1.7.4
bcrypt>=3.2.2